Add signature generation
