+++ docker/env.test.json
@@ -21,7 +21,8 @@
       "SUPPLIER_CASE_STUDY_URL",
       "SUPPLIER_PROFILE_LIST_URL",
       "SUPPLIER_PROFILE_URL",
-      "UTM_COOKIE_DOMAIN"
+      "UTM_COOKIE_DOMAIN",
+      "EXTERNAL_SECRET"
     ],
     "optional": [
       "PORT",
